Let ArtworkInfoShortcutController observe the wallpaper's lifecycle
Trigger updating the Artwork Info shortcut only when the wallpaper is active.

Also removes the need to have an initialize() method at all in MuzeiWallpaperService!
